# Optimized HexToBytes function